requirements.txt
pandas
pandas-gbq
tqdm
google-cloud-bigquery
google-api-python-client
requests
loguru